+++ apps/web/lib/db.ts
@@ -917,46 +917,3 @@ export async function getQuotaStatus(params: {
     resetAt: null,
   };
 }
-
-export type HostedSessionDeadline = {
-  sessionId: string;
-  taskSlug: string;
-  status: string;
-  sequenceIndex: number;
-  expiresAt: string | null;
-  timeLimitMinutes: number | null;
-};
-
-export async function listHostedSessionDeadlines(runId: string): Promise<HostedSessionDeadline[]> {
-  const supabase = getSupabase();
-
-  const { data, error } = await supabase
-    .from("hosted_web_sessions")
-    .select("id, task_slug, status, sequence_index, expires_at, metadata")
-    .eq("run_id", runId)
-    .order("sequence_index", { ascending: true });
-
-  if (error || !data) {
-    throw error ?? new Error("Failed to load hosted session deadlines");
-  }
-
-  return data.map((item) => {
-    const metadata =
-      typeof item.metadata === "object" && item.metadata !== null
-        ? (item.metadata as Record<string, unknown>)
-        : {};
-    const timeLimitMinutes =
-      typeof metadata.timeLimitMinutesPerTestcase === "number" && Number.isFinite(metadata.timeLimitMinutesPerTestcase)
-        ? metadata.timeLimitMinutesPerTestcase
-        : null;
-
-    return {
-      sessionId: item.id,
-      taskSlug: item.task_slug ?? "hosted-task",
-      status: item.status ?? "created",
-      sequenceIndex: item.sequence_index ?? 0,
-      expiresAt: item.expires_at ?? null,
-      timeLimitMinutes,
-    };
-  });
-}
